[WIP] first try: update can be done by Arachne, but cannot support pipelined requests.
The worker threads will create Arachne threads to handle SET request and send back the results.
The worker threads can directly read the next command.

However, there are some race conditions for c->item and other buffers.
Because memcached uses the same buffer for the same connection,
and we are using multiple concurrent Arachne threads.

So currently pipelining can cause segfault, because memcached will reuse the
same buffer to store new item and command.
We will need to change to multiple buffers to increase the parallelism.