Put connection dispatcher back to kernel thread
Since the connection dispatcher is not heavily loaded, we don't need an
exclusive arachne thread to deal with that. Also change event_base_loop flag
back to 0, blocking mode. Because the tight nonblock loop would be harmful to
performance at high load. We can change back to EVLOOP_NONBLOCK if we want to
guarantee the low latency at low load.